Sync device time on remote streaming start
If the playlist has remote streaming tracks and getting the next
streaming is delayed, the beginning of the next track is missing.
We synchronize the device time by `raopcl_flush` and `raopcl_stop`
since the time of the device is running while getting the next
streaming url.
Sync would be executed in the next `raopcl_accept_frames`.